Move shapes classes out of skyCatalogs.py into their own module; clean up some import statements
