+++ python/desc/skycatalogs/skyCatalogs.py
@@ -3,7 +3,6 @@ import sys
 import re
 import yaml
 import logging
-from collections import namedtuple
 import healpy
 import numpy as np
 import numpy.ma as ma
@@ -11,76 +10,14 @@ import pyarrow.parquet as pq
 from astropy import units as u
 from desc.skycatalogs.objects.base_object import load_lsst_bandpasses
 from desc.skycatalogs.objects.base_object import  CatalogContext
-from desc.skycatalogs.objects import *
-from desc.skycatalogs.readers import *
+from desc.skycatalogs.objects.base_object import  ObjectList, ObjectCollection
 from desc.skycatalogs.readers import ParquetReader
 from desc.skycatalogs.utils.sed_tools import ObservedSedFactory
 from desc.skycatalogs.utils.sed_tools import MilkyWayExtinction
 from desc.skycatalogs.utils.config_utils import Config
+from desc.skycatalogs.utils.shapes import Box, Disk, PolygonalRegion
 
-__all__ = ['SkyCatalog', 'open_catalog', 'Box', 'Disk', 'PolygonalRegion']
-
-
-Box = namedtuple('Box', ['ra_min', 'ra_max', 'dec_min', 'dec_max'])
-
-# radius is measured in arcseconds
-Disk = namedtuple('Disk', ['ra', 'dec', 'radius_as'])
-
-from lsst.sphgeom import ConvexPolygon, UnitVector3d, LonLat
-class PolygonalRegion:
-
-    def __init__(self, vertices_radec=None, convex_polygon=None):
-        '''
-        Supply either an object of type lsst.sphgeom.ConvexPolygon
-        or a list of vertices, each a tuple (ra, dec) in degrees,
-        which describe a convex polygon
-        '''
-        if convex_polygon:
-            if isinstance(convex_polygon, ConvexPolygon):
-                self._convex_polygon = convex_polygon
-                return
-        if vertices_radec:
-            if not isinstance(vertices_radec, list):
-                raise TypeError(f'PolygonalRegion: Argument {vertices_radec} is not a list')
-            vertices = [UnitVector3d(LonLat.fromDegrees(v_rd[0], v_rd[1])) for v_rd in vertices_radec]
-            self._convex_polygon = ConvexPolygon(vertices)
-            return
-        raise ValueError('PolygonalRegion: Either vertices_radec or convex_polygon must have an acceptable value')
-
-    def get_vertices(self):
-        '''
-        Return vertices as list of 3d vectors
-        '''
-        return self._convex_polygon.getVertices()
-
-    def get_vertices_radec(self):
-        v3d = self.get_vertices()
-        vertices_radec = []
-        for v in v3d:
-            vertices_radec.append((LonLat.longitudeOf(v).asDegrees(),
-                                   LonLat.latitudeOf(v).asDegrees()))
-        return vertices_radec
-
-    def get_containment_mask(self, ra, dec, included=True):
-        '''
-        Parameters
-        ----------
-        ra, dec      parallel float arrays, units are degrees. Together
-                     they define the list of points to be checked for
-                     containment
-        included     boolean   If true, mask bit will be set to True for
-                               contained points, else False.    Reverse
-                               the settings if included is False
-        '''
-        # convert to radians
-        ra = [(r * u.degree).to_value(u.radian) for r in ra]
-        dec = [(d * u.degree).to_value(u.radian) for d in dec]
-
-        mask = self._convex_polygon.contains(ra, dec)
-        if included:
-            return mask
-        else:
-            return np.logical_not(mask)
+__all__ = ['SkyCatalog', 'open_catalog']
 
 # This function should maybe be moved to utils
 def _get_intersecting_hps(hp_ordering, nside, region):
@@ -449,15 +386,6 @@ class SkyCatalog(object):
                                                       datetime=None)
             object_list.append_object_list(new_list)
 
-        # This now occurs inside get_object_type_by_region for
-        # object types partitioned by healpixel
-        # for hp in hps:
-        #     # Maybe have a multiprocessing switch? Run-time option when
-        #     # catalog is opened?
-        #     c = self.get_objects_by_hp(hp, region, obj_types, datetime)
-        #     if (len(c)) > 0:
-        #         object_list.append_object_list(c)
-
         return object_list
 
     def get_object_type_by_region(self, region, object_type, datetime=None):
@@ -525,7 +453,7 @@ class SkyCatalog(object):
         for f in f_list:
             if self._hp_info[hp]['files'][f] is None:            # no reader yet
                 full_path = os.path.join(self._cat_dir, f)
-                the_reader = parquet_reader.ParquetReader(full_path, mask=None)
+                the_reader = ParquetReader(full_path, mask=None)
                 self._hp_info[hp]['files'][f] = the_reader
             else:
                 the_reader = self._hp_info[hp]['files'][f]
