+++ python/desc/skycatalogs/utils/__init__.py
@@ -5,3 +5,4 @@ from .exceptions import *
 from .parquet_schema_utils import *
 from .sed_tools import *
 from .sn_tools import *
+from .shapes import *
